# Conflicts:
#	src/main/java/com/mercadolibre/w4g9projetofinal/controller/BuyerController.java
#	src/main/java/com/mercadolibre/w4g9projetofinal/service/RepresentativeService.java
#	src/main/java/com/mercadolibre/w4g9projetofinal/util/Instantiation.java
#	src/test/java/com/mercadolibre/w4g9projetofinal/test/unit/RepresentativeServiceTest.java

+++ src/main/java/com/mercadolibre/w4g9projetofinal/dtos/converter/BatchConverter.java
@@ -51,7 +51,7 @@ public class BatchConverter {
     }
     public static BatchResponseDTO convertEntityToDto (Batch b){
         return new BatchResponseDTO(
-                b.getId().intValue(),
+                b.getId(),
                 b.getAdvertise().getId(),
                 b.getCurrentTemperature(),
                 b.getMinTemperature(),
